+++ R/add_cumulative_yield.R
@@ -12,10 +12,10 @@
 
 #' @title Add daily cumulative runoff yield flows on an annual basis
 #'
-#' @description Add a column of daily cumulative runoff yield flows on an annual basis to a streamflow dataset. Adds the runoff yield 
-#'    discharge from each day with the previous day(s) for each year, in units of millimetres. Converts cumulative discharge to a 
-#'    depth of water based on the upstream drainge basin area.The cumulative flows restart every year and are only calculated in
-#'    years with complete data.
+#' @description Add a column of rolling daily cumulative runoff yield flows on an annual basis to a streamflow dataset. Adds the 
+#'    runoff yield discharge from each day with the previous day(s) for each year, in units of millimetres. Converts cumulative 
+#'    discharge to a depth of water based on the upstream drainge basin area.The cumulative flows restart every year and are only 
+#'    calculated in years with complete data.
 #'
 #' @inheritParams calc_annual_stats
 #' @inheritParams add_basin_area