+++ R/calc_daily_stats.R
@@ -10,7 +10,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
-#' @title Calculate the daily summary statistics
+#' @title Calculate daily summary statistics
 #'
 #' @description Calculates the daily mean, median, maximum, minimum, and percentiles for each day of the year of daily flow values 
 #'    from a streamflow dataset. Calculate the statistics from all daily discharge values from all years, unless specified. Can determine
@@ -186,7 +186,7 @@ calc_daily_stats <- function(data = NULL,
   
 
   # Recheck if station_number/grouping was in original flow_data and rename or remove as necessary
-  if("STATION_NUMBER" %in% orig_cols) {
+  if(as.character(substitute(groups)) %in% orig_cols) {
     names(daily_stats)[names(daily_stats) == "STATION_NUMBER"] <- as.character(substitute(groups))
   } else {
     daily_stats <- dplyr::select(daily_stats, -STATION_NUMBER)