+++ R/zzz.R
@@ -70,7 +70,11 @@
           "StatMonth",
           "Units",
           "WaterYear",
-          "WaterDayofYear")
+          "WaterDayofYear",
+          "Min_1_Day",
+          "Min_3_Day",
+          "Min_7_Day",
+          "Min_30_Day")
       )
     invisible()
   }